Adding old code for the router module
