+++ modules/router/src/tb_router_tlm.cpp
@@ -15,9 +15,9 @@ using namespace std;
 #include <tlm_utils/simple_target_socket.h>
 #include <tlm_utils/peq_with_cb_and_phase.h>
 
-#include "img_initiator.hpp"
-#include "img_target.hpp"
-#include "img_router.hpp"
+#include "img_initiator.cpp"
+#include "img_target.cpp"
+#include "img_router.cpp"
 
 #include "common_func.hpp"
 #include "important_defines.hpp"
