+++ modules/router/src/transaction_memory_manager.cpp
@@ -1,45 +1,81 @@
-#include "transaction_memory_manager.hpp"
+// **************************************************************************************
+// User-defined memory manager, which maintains a pool of transactions
+// From TLM Duolos tutorials
+// **************************************************************************************
+#ifndef TRANSACTION_MEMORY_MANAGER_CPP
+#define TRANSACTION_MEMORY_MANAGER_CPP
+
+#include <systemc.h>
 using namespace sc_core;
 using namespace sc_dt;
 using namespace std;
 
-gp_t *mm::allocate()
+#include <tlm.h>
+#include <tlm_utils/simple_initiator_socket.h>
+#include <tlm_utils/simple_target_socket.h>
+#include <tlm_utils/peq_with_cb_and_phase.h>
+
+class mm: public tlm::tlm_mm_interface
 {
-#ifdef DEBUG
-  cout << "----------------------------- Called allocate(), #trans = " << ++count << endl;
-#endif // DEBUG
-  gp_t *ptr;
+  typedef tlm::tlm_generic_payload gp_t;
 
-  if (free_list)
+public:
+  mm() : free_list(0), empties(0)
+  #ifdef DEBUG
+  , count(0)
+  #endif // DEBUG
+  {}
+
+  gp_t* allocate()
   {
-    ptr = free_list->trans;
-    empties = free_list;
-    free_list = free_list->next;
-  }
-  else
-  {
-    ptr = new gp_t(this);
-  }
-
-  return ptr;
-}
-
-void mm::free(gp_t *trans)
-{
-#ifdef DEBUG
-  cout << "----------------------------- Called free(), #trans = " << --count << endl;
-#endif // DEBUG
-  if (!empties)
-  {
-    empties = new access;
-    empties->next = free_list;
-    empties->prev = 0;
-
+    #ifdef DEBUG
+      cout << "----------------------------- Called allocate(), #trans = " << ++count << endl;
+    #endif // DEBUG
+    gp_t* ptr;
     if (free_list)
-      free_list->prev = empties;
+    {
+      ptr = free_list->trans;
+      empties = free_list;
+      free_list = free_list->next;
+    }
+    else
+    {
+      ptr = new gp_t(this);
+    }
+    return ptr;
   }
 
-  free_list = empties;
-  free_list->trans = trans;
-  empties = free_list->prev;
-}
+  void  free(gp_t* trans)
+  {
+    #ifdef DEBUG
+      cout << "----------------------------- Called free(), #trans = " << --count << endl;
+    #endif // DEBUG
+    if (!empties)
+    {
+      empties = new access;
+      empties->next = free_list;
+      empties->prev = 0;
+      if (free_list)
+        free_list->prev = empties;
+    }
+    free_list = empties;
+    free_list->trans = trans;
+    empties = free_list->prev;
+  }
+
+private:
+  struct access
+  {
+    gp_t* trans;
+    access* next;
+    access* prev;
+  };
+
+  access* free_list;
+  access* empties;
+
+  #ifdef DEBUG
+  int     count;
+  #endif // DEBUG
+};
+#endif // TRANSACTION_MEMORY_MANAGER_CPP
